# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PetHub Import Export Service
  version: 1.0.0
  description: Versioned, scoped and auditable CSV/XLSX import batches and purpose-bound exports.
servers: [{url: /}]
security: [{bearerAuth: []}]
paths:
  /internal/import-export/batches:
    post:
      operationId: createImportBatch
      parameters: [{$ref: '#/components/parameters/OrganizationId'}, {$ref: '#/components/parameters/UnitId'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CorrelationId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateBatchRequest'}}}}
      responses: {'202': {description: Queued, content: {application/json: {schema: {$ref: '#/components/schemas/Batch'}}}}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '409': {$ref: '#/components/responses/Conflict'}, '413': {$ref: '#/components/responses/TooLarge'}, '422': {$ref: '#/components/responses/Invalid'}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {$ref: '#/components/responses/Unavailable'}}
    get:
      operationId: listImportBatches
      parameters: [{$ref: '#/components/parameters/OrganizationId'}, {$ref: '#/components/parameters/CorrelationId'}]
      responses: {'200': {description: Scoped batches, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Batch'}}}}}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}}
  /internal/import-export/batches/{id}:
    get:
      operationId: getImportBatch
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/OrganizationId'}]
      responses: {'200': {description: Batch, content: {application/json: {schema: {$ref: '#/components/schemas/Batch'}}}}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}}
  /internal/import-export/batches/{id}/run:
    post:
      operationId: runImportBatch
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/OrganizationId'}]
      responses: {'200': {description: Preview completed, content: {application/json: {schema: {$ref: '#/components/schemas/Batch'}}}}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '413': {$ref: '#/components/responses/TooLarge'}, '422': {$ref: '#/components/responses/Invalid'}}
  /internal/import-export/batches/{id}/approve:
    post:
      operationId: approveImportBatch
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/OrganizationId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ApproveBatchRequest'}}}}
      responses: {'200': {description: Approved, content: {application/json: {schema: {$ref: '#/components/schemas/Batch'}}}}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/Invalid'}}
  /internal/import-export/batches/{id}/rollback:
    post:
      operationId: rollbackImportBatch
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/OrganizationId'}]
      responses: {'200': {description: Logical rollback, content: {application/json: {schema: {$ref: '#/components/schemas/Batch'}}}}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}}
  /internal/import-export/exports:
    post:
      operationId: createExport
      parameters: [{$ref: '#/components/parameters/OrganizationId'}, {$ref: '#/components/parameters/UnitId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ExportRequest'}}}}
      responses: {'202': {description: Export ready with expiring signed URL, content: {application/json: {schema: {$ref: '#/components/schemas/ExportResult'}}}}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/Invalid'}, '503': {$ref: '#/components/responses/Unavailable'}}
  /internal/import-export/exports/{id}/download:
    get:
      operationId: downloadExport
      parameters: [{$ref: '#/components/parameters/Id'}, {name: X-Download-Token, in: header, required: true, schema: {type: string}}, {$ref: '#/components/parameters/OrganizationId'}]
      responses: {'200': {description: CSV export}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}}
components:
  securitySchemes: {bearerAuth: {type: http, scheme: bearer}}
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    OrganizationId: {name: X-Organization-Id, in: header, required: true, schema: {type: string, format: uuid}}
    UnitId: {name: X-Unit-Id, in: header, required: false, schema: {type: string, format: uuid}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, maxLength: 160}}
    CorrelationId: {name: X-Correlation-Id, in: header, required: false, schema: {type: string, format: uuid}}
  schemas:
    ManifestColumn: {type: object, required: [name, type], properties: {name: {type: string}, type: {type: string, enum: [STRING, EMAIL, PHONE, CPF, CNPJ, EAN, CURRENCY, DATE, TIMEZONE, UUID, UNIT_ID]}, required: {type: boolean}, unique: {type: boolean}}}
    ImportManifest: {type: object, required: [manifestId, version, purpose, format, encoding, timezone, scopePolicy, columns], properties: {manifestId: {type: string}, version: {type: integer, minimum: 1}, purpose: {type: string}, format: {type: string, enum: [CSV, XLSX]}, encoding: {type: string, example: UTF-8}, timezone: {type: string, example: America/Sao_Paulo}, scopePolicy: {type: string, enum: [ORGANIZATION_SHARED, UNIT_LINKED]}, organizationId: {type: string, format: uuid}, unitId: {type: string, format: uuid}, columns: {type: array, minItems: 1, items: {$ref: '#/components/schemas/ManifestColumn'}}}}
    CreateBatchRequest: {type: object, required: [manifest, fileName, contentType, encoding, checksum, uploadSignature, content], properties: {manifest: {$ref: '#/components/schemas/ImportManifest'}, fileName: {type: string}, contentType: {type: string}, encoding: {type: string}, checksum: {type: string, pattern: '^[a-fA-F0-9]{64}$'}, uploadSignature: {type: string}, content: {type: string}, dryRun: {type: boolean, default: true}}}
    ApproveBatchRequest: {type: object, required: [approvalReason, checksum], properties: {approvalReason: {type: string}, checksum: {type: string}}}
    Batch: {type: object, required: [id, organizationId, manifestId, manifestVersion, status, checksum, checkpoint], properties: {id: {type: string, format: uuid}, organizationId: {type: string, format: uuid}, unitId: {type: string, format: uuid}, manifestId: {type: string}, manifestVersion: {type: integer}, status: {type: string, enum: [queued, running, failed, succeeded, approved, rolled_back]}, dryRun: {type: boolean}, checksum: {type: string}, rows: {type: integer}, validRows: {type: integer}, invalidRows: {type: integer}, conflicts: {type: integer}, checkpoint: {type: string}, preview: {type: object}}}
    ExportRequest: {type: object, required: [purpose, format, organizationId, timezone], properties: {purpose: {type: string}, format: {type: string, enum: [CSV, XLSX]}, organizationId: {type: string, format: uuid}, unitId: {type: string, format: uuid}, rows: {type: array, items: {type: object, additionalProperties: {type: string}}}, includePii: {type: boolean, default: false}, timezone: {type: string}}}
    ExportResult: {type: object, required: [exportId, purpose, format, downloadUrl, checksum, expiresAt, piiMasked], properties: {exportId: {type: string, format: uuid}, purpose: {type: string}, format: {type: string}, downloadUrl: {type: string}, checksum: {type: string}, expiresAt: {type: string, format: date-time}, piiMasked: {type: boolean}}}
    ProblemDetails: {type: object, required: [status, code, message, correlationId, timestamp], properties: {status: {type: integer}, code: {type: string}, message: {type: string}, details: {type: object}, correlationId: {type: string}, timestamp: {type: string, format: date-time}}}
  responses:
    BadRequest: {description: Bad request, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemDetails'}}}}
    Unauthorized: {description: Authentication/signature failure, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemDetails'}}}}
    Forbidden: {description: Cross-scope or missing scope, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemDetails'}}}}
    NotFound: {description: Outside organization scope, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemDetails'}}}}
    Conflict: {description: Idempotency/state/checksum/expiry conflict, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemDetails'}}}}
    TooLarge: {description: Size or row limit exceeded, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemDetails'}}}}
    Invalid: {description: Manifest/encoding/row validation error, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemDetails'}}}}
    RateLimited: {description: Rate limit, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemDetails'}}}}
    Unavailable: {description: Dependency unavailable, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemDetails'}}}}